Change MusicFadeEnum type to ushort
* We read a ushort to get the MusicFade value anyways, so the enum should probably be the same type

+++ Libellus Library/Event/Types/Frame/PmdTarget_Bgm.cs
@@ -18,7 +18,7 @@ namespace LibellusLibrary.Event.Types.Frame
 
 		// This is only accurate for P3; P4(G) removes some of these from the Event Editor menu
 		// TODO: Figure out a way to address that discrepancy
-		internal enum MusicFadeEnum : byte
+		internal enum MusicFadeEnum : ushort
 		{
 			PLAY = 0,
 			FADEIN = 1,
